--- src/a2a/utils/artifact.py
@@ -1,92 +1,0 @@
-"""Utility functions for creating A2A Artifact objects."""
-
-import uuid
-
-from typing import Any
-
-from google.protobuf.struct_pb2 import Struct, Value
-
-from a2a.types.a2a_pb2 import Artifact, Part
-from a2a.utils.parts import get_text_parts
-
-
-def new_artifact(
-    parts: list[Part],
-    name: str,
-    description: str | None = None,
-) -> Artifact:
-    """Creates a new Artifact object.
-
-    Args:
-        parts: The list of `Part` objects forming the artifact's content.
-        name: The human-readable name of the artifact.
-        description: An optional description of the artifact.
-
-    Returns:
-        A new `Artifact` object with a generated artifact_id.
-    """
-    return Artifact(
-        artifact_id=str(uuid.uuid4()),
-        parts=parts,
-        name=name,
-        description=description,
-    )
-
-
-def new_text_artifact(
-    name: str,
-    text: str,
-    description: str | None = None,
-) -> Artifact:
-    """Creates a new Artifact object containing only a single text Part.
-
-    Args:
-        name: The human-readable name of the artifact.
-        text: The text content of the artifact.
-        description: An optional description of the artifact.
-
-    Returns:
-        A new `Artifact` object with a generated artifact_id.
-    """
-    return new_artifact(
-        [Part(text=text)],
-        name,
-        description,
-    )
-
-
-def new_data_artifact(
-    name: str,
-    data: dict[str, Any],
-    description: str | None = None,
-) -> Artifact:
-    """Creates a new Artifact object containing only a single data Part.
-
-    Args:
-        name: The human-readable name of the artifact.
-        data: The structured data content of the artifact.
-        description: An optional description of the artifact.
-
-    Returns:
-        A new `Artifact` object with a generated artifact_id.
-    """
-    struct_data = Struct()
-    struct_data.update(data)
-    return new_artifact(
-        [Part(data=Value(struct_value=struct_data))],
-        name,
-        description,
-    )
-
-
-def get_artifact_text(artifact: Artifact, delimiter: str = '\n') -> str:
-    """Extracts and joins all text content from an Artifact's parts.
-
-    Args:
-        artifact: The `Artifact` object.
-        delimiter: The string to use when joining text from multiple TextParts.
-
-    Returns:
-        A single string containing all text content, or an empty string if no text parts are found.
-    """
-    return delimiter.join(get_text_parts(artifact.parts))
