+++ tests/client/transports/test_rest_client.py
@@ -8,7 +8,7 @@ from google.protobuf import json_format
 from google.protobuf.timestamp_pb2 import Timestamp
 from httpx_sse import EventSource, ServerSentEvent
 
-from a2a.client import create_text_message_object
+from a2a.helpers.types import new_text_message
 from a2a.client.client import ClientCallContext
 from a2a.client.errors import A2AClientError
 from a2a.client.transports.rest import RestTransport
@@ -83,7 +83,7 @@ class TestRestTransport:
             url='http://agent.example.com/api',
         )
         params = SendMessageRequest(
-            message=create_text_message_object(content='Hello stream')
+            message=new_text_message(text='Hello stream')
         )
         mock_event_source = AsyncMock(spec=EventSource)
         mock_event_source.response = MagicMock(spec=httpx.Response)
@@ -121,7 +121,7 @@ class TestRestTransport:
             url='http://agent.example.com/api',
         )
         params = SendMessageRequest(
-            message=create_text_message_object(content='Hello')
+            message=new_text_message(text='Hello')
         )
 
         mock_build_request = MagicMock(
@@ -173,7 +173,7 @@ class TestRestTransport:
             url='http://agent.example.com/api',
         )
         params = SendMessageRequest(
-            message=create_text_message_object(content='Hello')
+            message=new_text_message(text='Hello')
         )
         context = ClientCallContext(timeout=10.0)
 
@@ -247,7 +247,7 @@ class TestRestTransportExtensions:
             url='http://agent.example.com/api',
         )
         params = SendMessageRequest(
-            message=create_text_message_object(content='Hello')
+            message=new_text_message(text='Hello')
         )
 
         # Mock the build_request method to capture its inputs
@@ -294,7 +294,7 @@ class TestRestTransportExtensions:
             url='http://agent.example.com/api',
         )
         params = SendMessageRequest(
-            message=create_text_message_object(content='Hello stream')
+            message=new_text_message(text='Hello stream')
         )
 
         mock_event_source = AsyncMock(spec=EventSource)
@@ -343,7 +343,7 @@ class TestRestTransportExtensions:
             url='http://agent.example.com/api',
         )
         request = SendMessageRequest(
-            message=create_text_message_object(content='Error stream')
+            message=new_text_message(text='Error stream')
         )
 
         mock_event_source = AsyncMock(spec=EventSource)
@@ -524,7 +524,7 @@ class TestRestTransportTenant:
                 'send_message',
                 SendMessageRequest(
                     tenant='my-tenant',
-                    message=create_text_message_object(content='hi'),
+                    message=new_text_message(text='hi'),
                 ),
                 '/my-tenant/message:send',
             ),
@@ -686,7 +686,7 @@ class TestRestTransportTenant:
                 'send_message_streaming',
                 SendMessageRequest(
                     tenant='my-tenant',
-                    message=create_text_message_object(content='hi'),
+                    message=new_text_message(text='hi'),
                 ),
                 '/my-tenant/message:stream',
             ),
