+++ tests/integration/test_end_to_end.py
@@ -37,7 +37,8 @@ from a2a.types import (
     TaskState,
     a2a_pb2_grpc,
 )
-from a2a.utils import TransportProtocol, new_task
+from a2a.utils import TransportProtocol
+from a2a.helpers.types import new_task_from_request
 from a2a.utils.errors import InvalidParamsError
 
 
@@ -101,7 +102,7 @@ class MockAgentExecutor(AgentExecutor):
         # Task-based response.
         task = context.current_task
         if not task:
-            task = new_task(context.message)
+            task = new_task_from_request(context.message)
             await event_queue.enqueue_event(task)
 
         task_updater = TaskUpdater(
