+++ tests/server/request_handlers/test_default_request_handler_v2.py
@@ -54,7 +54,7 @@ from a2a.types.a2a_pb2 import (
     TaskState,
     TaskStatus,
 )
-from a2a.utils import new_agent_text_message, new_task
+from a2a.helpers.types import new_text_message, new_task_from_request
 
 
 def create_default_agent_card():
@@ -211,8 +211,8 @@ async def test_on_list_tasks_applies_history_length():
     """Test on_list_tasks applies history length filter."""
     mock_task_store = AsyncMock(spec=TaskStore)
     history = [
-        new_agent_text_message('Hello 1!'),
-        new_agent_text_message('Hello 2!'),
+        new_text_message('Hello 1!'),
+        new_text_message('Hello 2!'),
     ]
     task2 = create_sample_task(task_id='task2')
     task2.history.extend(history)
@@ -274,7 +274,7 @@ class HelloAgentExecutor(AgentExecutor):
             assert context.message is not None, (
                 'A message is required to create a new task'
             )
-            task = new_task(context.message)
+            task = new_task_from_request(context.message)
             await event_queue.enqueue_event(task)
         updater = TaskUpdater(event_queue, task.id, task.context_id)
         try:
