+++ tests/utils/test_agent_card_display.py
@@ -9,7 +9,7 @@ from a2a.types.a2a_pb2 import (
     AgentProvider,
     AgentSkill,
 )
-from a2a.utils.agent_card import display_agent_card
+from a2a.helpers.agent_card import display_agent_card
 
 
 @pytest.fixture
